Implemented on-the-fly ed

+++ NetKet/Hamiltonian/MatrixWrapper/direct_matrix_wrapper.hpp
@@ -42,7 +42,7 @@ class DirectMatrixWrapper : public AbstractMatrixWrapper<WfType> {
     WfType result(dim_);
     result.setZero();
 
-    for (int i = 0; i < dim_; ++i) {
+    for (size_t i = 0; i < dim_; ++i) {
       auto v = hilbert_index.NumberToState(i);
 
       std::vector<std::complex<double>> matrix_elements;
